compose: first commit, really the third version
This set of scripts generates an XCompose file from the latest official
list of Unicode codepoints and a set of tables of predefined mappings.
The goal is to avoid manually creating mappings of character groups.
Plus it nicely formats and aligns the XCompose file.

First came the shell script, then came the generator loops, and finally
the tables of additional characters. This is not complete: I'm bad at
Python, the code needs a bit of refactoring, the Unicode versions do not
match (except with Python 3.5), a shell script is the poorest man's form
of metaprogramming, etc. Plus some of the tables are just plain wrong.

But it is good enough for now.