add USB port switching server
This is useful for KVM/qemu virtual machines with USB controllers
passed to them. It allows you to control which controller certain ports
are attached to by sending a message over a TCP socket.

If this runs as root, it allows control over USB ports without
privilege escalation. Additionally, if an emulated serial port is
hooked up to the server, it can be controlled from the virtual machine.